requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
celery==5.3.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
django==4.2
django-celery-beat==2.9.0
django-cors-headers==3.13.0
django-timezone-field==7.2.1
djangorestframework==3.14.0
kombu==5.6.2
packaging==26.0
prompt-toolkit==3.0.52
psycopg2-binary==2.9.6
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2026.1.post1
redis==4.5.1
six==1.17.0
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2025.3
vine==5.1.0
wcwidth==0.6.0
